// shared/enums.ts
export enum ProductType {
  ADULT = "adult",
  CHILD = "child",
  FAMILY = "family",
  FAST_PASS = "fast_pass",
  VOUCHER = "voucher",
}

export enum ProductCategory {
  ADMISSION = "admission",
  SERVICE = "service",
}

export enum ProductTag {
  BESTSELLER = "bestseller",
  PROMOTION = "promotion",
  LIMITED = "limited",
}